+++ tasks/seq/Sadikov_I_Gauss_Linear_Filtration/include/seq_task.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cmath>
 #include <cstdlib>
 #include <ctime>
 #include <iostream>
@@ -27,7 +28,7 @@ class LinearFiltration : public ppc::core::Task {
   bool post_processing() override;
   void CalculateGaussMatrix();
   void CalculateNewPixelValue(int iIndex, int jIndex);
-  bool CheckIndex(int index) const noexcept { return index >= 0 and index < m_pixelsMatrix.size(); }
+  bool CheckIndex(int index) const noexcept { return index >= 0 and index < static_cast<int>(m_pixelsMatrix.size()); }
   bool CheckRowIndex(int index) const noexcept { return index >= 0 and index < m_rowsCount; }
   bool CheckColumnIndex(int index) const noexcept { return index >= 0 and index < m_columnsCount; }
 };
